Updated integration tests
