+++ magpie-aws/src/test/java/io/openraven/magpie/plugins/aws/discovery/services/DynamoDbDiscoveryIT.java
@@ -4,6 +4,7 @@ import io.openraven.magpie.api.Emitter;
 import io.openraven.magpie.api.MagpieEnvelope;
 import io.openraven.magpie.plugins.aws.discovery.AWSUtils;
 import io.openraven.magpie.plugins.aws.discovery.BackupUtils;
+import io.openraven.magpie.plugins.aws.discovery.ClientCreators;
 import io.openraven.magpie.plugins.aws.discovery.services.base.BaseAWSServiceIT;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -49,7 +50,7 @@ class DynamoDbDiscoveryIT extends BaseAWSServiceIT {
   @Test
   public void testDynamoDBDiscovery() {
     // given
-    DynamoDbClient dynamoDbClient = AWSUtils.configure(DynamoDbClient.builder(), BASE_REGION);
+    DynamoDbClient dynamoDbClient = ClientCreators.localClientCreator(BASE_REGION).apply(DynamoDbClient.builder()).build();
     BackupUtils.init(BASE_REGION, backupClient);
     when(backupClient.listBackupJobs(any(ListBackupJobsRequest.class)))
       .thenReturn(ListBackupJobsResponse.builder().backupJobs(Collections.emptyList()).build());
@@ -61,7 +62,9 @@ class DynamoDbDiscoveryIT extends BaseAWSServiceIT {
       BASE_REGION,
       emitter,
       dynamoDbClient,
-      ACCOUNT);
+      ACCOUNT,
+      ClientCreators.localClientCreator(BASE_REGION)
+    );
 
     // then
     Mockito.verify(emitter).emit(envelopeCapture.capture());
