+++ magpie-aws/src/test/java/io/openraven/magpie/plugins/aws/discovery/services/Route53DiscoveryIT.java
@@ -2,6 +2,7 @@ package io.openraven.magpie.plugins.aws.discovery.services;
 
 import io.openraven.magpie.api.Emitter;
 import io.openraven.magpie.api.MagpieEnvelope;
+import io.openraven.magpie.plugins.aws.discovery.ClientCreators;
 import io.openraven.magpie.plugins.aws.discovery.services.base.BaseAWSServiceIT;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -41,7 +42,8 @@ public class Route53DiscoveryIT extends BaseAWSServiceIT {
       BASE_REGION,
       emitter,
       LOGGER,
-      ACCOUNT
+      ACCOUNT,
+      ClientCreators.localClientCreator(BASE_REGION)
     );
 
     // then
