+++ magpie-aws/src/test/java/io/openraven/magpie/plugins/aws/discovery/services/base/BaseIAMServiceIT.java
@@ -1,12 +1,13 @@
 package io.openraven.magpie.plugins.aws.discovery.services.base;
 
 import io.openraven.magpie.plugins.aws.discovery.AWSUtils;
+import io.openraven.magpie.plugins.aws.discovery.ClientCreators;
 import software.amazon.awssdk.services.iam.IamClient;
 import software.amazon.awssdk.services.iam.model.PolicyScopeType;
 
 public abstract class BaseIAMServiceIT extends BaseAWSServiceIT {
 
-  protected static final IamClient IAMCLIENT = AWSUtils.configure(IamClient.builder(), BASE_REGION);
+  protected static final IamClient IAMCLIENT = ClientCreators.localClientCreator(BASE_REGION).apply((IamClient.builder())).build();
 
   protected static void removePolicies() {
     IAMCLIENT.listPolicies(req -> req.scope(PolicyScopeType.LOCAL))
