+++ magpie-aws/src/test/java/io/openraven/magpie/plugins/aws/discovery/services/disabled/RedshiftDiscoveryIT.java
@@ -2,6 +2,7 @@ package io.openraven.magpie.plugins.aws.discovery.services.disabled;
 
 import io.openraven.magpie.api.Emitter;
 import io.openraven.magpie.api.MagpieEnvelope;
+import io.openraven.magpie.plugins.aws.discovery.ClientCreators;
 import io.openraven.magpie.plugins.aws.discovery.services.RedshiftDiscovery;
 import io.openraven.magpie.plugins.aws.discovery.services.base.BaseAWSServiceIT;
 import org.junit.jupiter.api.BeforeAll;
@@ -43,7 +44,8 @@ public class RedshiftDiscoveryIT extends BaseAWSServiceIT {
       BASE_REGION,
       emitter,
       LOGGER,
-      ACCOUNT
+      ACCOUNT,
+      ClientCreators.localClientCreator(BASE_REGION)
     );
 
     // then
